+++ examples/apps/react-todomvc-ts/src/store/selectors.ts
@@ -1,4 +1,4 @@
-import { createSelector } from "@inglorious/store/select"
+import { compute } from "@inglorious/store/select"
 
 import type { Filter, Task, AppState } from "../../types"
 
@@ -6,17 +6,17 @@ export const selectValue = (entities: AppState) => entities.form.value
 export const selectTasks = (entities: AppState) => entities.list.tasks
 
 export const selectTasksCount = (filter?: Filter) =>
-  createSelector(
-    [selectTasks],
+  compute<AppState, [Task[]], number>(
     (tasks: Task[]) => getTasks(tasks, filter).length,
+    [selectTasks],
   )
 
 export const selectActiveFilter = (entities: AppState): Filter =>
   entities.footer.activeFilter
 
-export const selectFilteredTasks = createSelector(
-  [selectTasks, selectActiveFilter],
+export const selectFilteredTasks = compute<AppState, [Task[], Filter], Task[]>(
   (tasks: Task[], activeFilter: Filter) => getTasks(tasks, activeFilter),
+  [selectTasks, selectActiveFilter],
 )
 
 function getTasks(tasks: Task[], filter?: Filter) {