+++ examples/apps/react-todomvc-ts/CHANGELOG.md
@@ -1,5 +1,13 @@
 # todomvc
 
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/store@9.1.0
+  - @inglorious/react-store@12.0.0
+
 ## 0.1.17
 
 ### Patch Changes