+++ examples/apps/web-todomvc-cs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # lit-todomvc
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@4.1.0
+
 ## 0.0.19
 
 ### Patch Changes