+++ packages/logo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inglorious/logo
 
+## 2.0.22
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/web@4.1.0
+
 ## 2.0.21
 
 ### Patch Changes